requirements.txt
streamlit==1.28.1
pandas==2.1.4
numpy==1.26.3
plotly==5.17.0
scikit-learn==1.3.2
mlflow==2.10.2
dagshub==0.3.44
joblib==1.3.2
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
flask==3.1.2
joblib==1.3.2
mlflow==2.12.1
requests==2.31.0
blinker==1.9.0
click==8.1.8
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
setuptools==68.0.0
gunicorn==20.1.0